Domain id 자료형 수정
Domain의 각 id를 int형에서 Long형으로 자료형 변환

+++ src/main/java/com/ulmu/dadjoketest/domain/Comment.java
@@ -11,23 +11,23 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id", unique = true, nullable = false)
-    private int commentId; //pk
+    private Long commentId; //pk
     @Column(name = "comment_detail", unique = false, nullable = false)
     private String commentDetail;  //댓글내용
     @Column(name = "created_at", unique = false, nullable = false)
     @CreationTimestamp
     private Date createdAt;    //댓글 작성일
     @Column(name = "great", unique = false, nullable = false)
-    private int great;  //댓글 좋아요 수
+    private Long great;  //댓글 좋아요 수
     @Column(name = "joke_id", unique = true, nullable = false)
-    private int jokeId; //댓글을 단 농담 id
+    private Long jokeId; //댓글을 단 농담 id
     @Column(name = "create_user_id", unique = true, nullable = false)
-    private int createUserUd; //댓글 작성자 id
+    private Long createUserUd; //댓글 작성자 id
     @Column(name = "comment_great_user_list", unique = false, nullable = false)
     private String greatUserIDList;     //좋아요 누른 유저 id list - 띄어쓰기로 구분해서 데이터 입력
                                         //DTO에서 Mapper가 List<String>을 ','로 구분되는 String으로 변환하여 통으로 저장하도록
 
-    public Comment(int commentId, String commentDetail, Date createdAt, int great, int jokeId, int createUserUd) {
+    public Comment(Long commentId, String commentDetail, Date createdAt, Long great, Long jokeId, Long createUserUd) {
         this.commentId = commentId;
         this.commentDetail = commentDetail;
         this.createdAt = createdAt;
@@ -36,11 +36,11 @@ public class Comment {
         this.createUserUd = createUserUd;
     }
 
-    public int getCommentId() {
+    public Long getCommentId() {
         return commentId;
     }
 
-    public void setCommentId(int commentId) {
+    public void setCommentId(Long commentId) {
         this.commentId = commentId;
     }
 
@@ -60,27 +60,27 @@ public class Comment {
         this.createdAt = createdAt;
     }
 
-    public int getGreat() {
+    public Long getGreat() {
         return great;
     }
 
-    public void setGreat(int great) {
+    public void setGreat(Long great) {
         this.great = great;
     }
 
-    public int getJokeId() {
+    public Long getJokeId() {
         return jokeId;
     }
 
-    public void setJokeId(int jokeId) {
+    public void setJokeId(Long jokeId) {
         this.jokeId = jokeId;
     }
 
-    public int getCreateUserUd() {
+    public Long getCreateUserUd() {
         return createUserUd;
     }
 
-    public void setCreateUserUd(int createUserUd) {
+    public void setCreateUserUd(Long createUserUd) {
         this.createUserUd = createUserUd;
     }
